Trait sgx_trts::ascii::AsciiExt

pub trait AsciiExt {
    type Owned;
    fn is_ascii(&self) -> bool;

    fn to_ascii_uppercase(&self) -> Self::Owned;

    fn to_ascii_lowercase(&self) -> Self::Owned;

    fn eq_ignore_ascii_case(&self, other: &Self) -> bool;

    fn make_ascii_uppercase(&mut self);

    fn make_ascii_lowercase(&mut self);

    fn is_ascii_alphabetic(&self) -> bool { ... }

    fn is_ascii_uppercase(&self) -> bool { ... }

    fn is_ascii_lowercase(&self) -> bool { ... }

    fn is_ascii_alphanumeric(&self) -> bool { ... }

    fn is_ascii_digit(&self) -> bool { ... }

    fn is_ascii_hexdigit(&self) -> bool { ... }

    fn is_ascii_punctuation(&self) -> bool { ... }

    fn is_ascii_graphic(&self) -> bool { ... }

    fn is_ascii_whitespace(&self) -> bool { ... }

    fn is_ascii_control(&self) -> bool { ... }
}

Associated Types

type Owned

Container type for copied ASCII characters.

Required Methods

fn is_ascii(&self) -> bool

Checks if the value is within the ASCII range.

fn to_ascii_uppercase(&self) -> Self::Owned

Makes a copy of the value in its ASCII upper case equivalent.

ASCII letters 'a' to 'z' are mapped to 'A' to 'Z', but non-ASCII letters are unchanged.

To uppercase the value in-place, use [make_ascii_uppercase].

To uppercase ASCII characters in addition to non-ASCII characters, use [str::to_uppercase].

fn to_ascii_lowercase(&self) -> Self::Owned

Makes a copy of the value in its ASCII lower case equivalent.

ASCII letters 'A' to 'Z' are mapped to 'a' to 'z', but non-ASCII letters are unchanged.

To lowercase the value in-place, use [make_ascii_lowercase].

To lowercase ASCII characters in addition to non-ASCII characters, use [str::to_lowercase].

fn eq_ignore_ascii_case(&self, other: &Self) -> bool

Checks that two values are an ASCII case-insensitive match.

Same as to_ascii_lowercase(a) == to_ascii_lowercase(b), but without allocating and copying temporaries.

fn make_ascii_uppercase(&mut self)

Converts this type to its ASCII upper case equivalent in-place.

ASCII letters 'a' to 'z' are mapped to 'A' to 'Z', but non-ASCII letters are unchanged.

To return a new uppercased value without modifying the existing one, use [to_ascii_uppercase].

fn make_ascii_lowercase(&mut self)

Converts this type to its ASCII lower case equivalent in-place.

ASCII letters 'A' to 'Z' are mapped to 'a' to 'z', but non-ASCII letters are unchanged.

To return a new lowercased value without modifying the existing one, use [to_ascii_lowercase].

Provided Methods

fn is_ascii_alphabetic(&self) -> bool

Checks if the value is an ASCII alphabetic character: U+0041 'A' ... U+005A 'Z' or U+0061 'a' ... U+007A 'z'. For strings, true if all characters in the string are ASCII alphabetic.

fn is_ascii_uppercase(&self) -> bool

Checks if the value is an ASCII uppercase character: U+0041 'A' ... U+005A 'Z'. For strings, true if all characters in the string are ASCII uppercase.

fn is_ascii_lowercase(&self) -> bool

Checks if the value is an ASCII lowercase character: U+0061 'a' ... U+007A 'z'. For strings, true if all characters in the string are ASCII lowercase.

fn is_ascii_alphanumeric(&self) -> bool

Checks if the value is an ASCII alphanumeric character: U+0041 'A' ... U+005A 'Z', U+0061 'a' ... U+007A 'z', or U+0030 '0' ... U+0039 '9'. For strings, true if all characters in the string are ASCII alphanumeric.

fn is_ascii_digit(&self) -> bool

Checks if the value is an ASCII decimal digit: U+0030 '0' ... U+0039 '9'. For strings, true if all characters in the string are ASCII digits.

fn is_ascii_hexdigit(&self) -> bool

Checks if the value is an ASCII hexadecimal digit: U+0030 '0' ... U+0039 '9', U+0041 'A' ... U+0046 'F', or U+0061 'a' ... U+0066 'f'. For strings, true if all characters in the string are ASCII hex digits.

fn is_ascii_punctuation(&self) -> bool

Checks if the value is an ASCII punctuation character: U+0021 ... U+002F ! " # $ % & ' ( ) * + , - . / U+003A ... U+0040 : ; < = > ? @ U+005B ... U+0060 [ \\ ] ^ _ \U+007B ... U+007E{ | } ~` For strings, true if all characters in the string are ASCII punctuation.

fn is_ascii_graphic(&self) -> bool

Checks if the value is an ASCII graphic character: U+0021 '@' ... U+007E '~'. For strings, true if all characters in the string are ASCII punctuation.

fn is_ascii_whitespace(&self) -> bool

Checks if the value is an ASCII whitespace character: U+0020 SPACE, U+0009 HORIZONTAL TAB, U+000A LINE FEED, U+000C FORM FEED, or U+000D CARRIAGE RETURN. For strings, true if all characters in the string are ASCII whitespace.

Rust uses the WhatWG Infra Standard's [definition of ASCII whitespace][infra-aw]. There are several other definitions in wide use. For instance, [the POSIX locale][pct] includes U+000B VERTICAL TAB as well as all the above characters, but—from the very same specification—[the default rule for "field splitting" in the Bourne shell][bfs] considers only SPACE, HORIZONTAL TAB, and LINE FEED as whitespace.

If you are writing a program that will process an existing file format, check what that format's definition of whitespace is before using this function.

fn is_ascii_control(&self) -> bool

Checks if the value is an ASCII control character: U+0000 NUL ... U+001F UNIT SEPARATOR, or U+007F DELETE. Note that most ASCII whitespace characters are control characters, but SPACE is not.