Struct sgx_trts::c_str::CString

pub struct CString { /* fields omitted */ }

A type representing an owned C-compatible string

Methods

impl CString

pub fn new<T: Into<Vec<u8>>>(t: T) -> Result<CString, NulError>

Creates a new C-compatible string from a container of bytes.

This method will consume the provided data and use the underlying bytes to construct a new string, ensuring that there is a trailing 0 byte.

Errors

This function will return an error if the bytes yielded contain an internal 0 byte. The error returned will contain the bytes as well as the position of the nul byte.

pub unsafe fn from_vec_unchecked(v: Vec<u8>) -> CString

Creates a C-compatible string from a byte vector without checking for interior 0 bytes.

This method is equivalent to new except that no runtime assertion is made that v contains no 0 bytes, and it requires an actual byte vector, not anything that can be converted to one with Into.

pub unsafe fn from_raw(ptr: *mut c_char) -> CString

Retakes ownership of a CString that was transferred to C.

Additionally, the length of the string will be recalculated from the pointer.

Safety

This should only ever be called with a pointer that was earlier obtained by calling into_raw on a CString. Other usage (e.g. trying to take ownership of a string that was allocated by foreign code) is likely to lead to undefined behavior or allocator corruption.

pub fn into_raw(self) -> *mut c_char

Transfers ownership of the string to a C caller.

The pointer must be returned to Rust and reconstituted using from_raw to be properly deallocated. Specifically, one should not use the standard C free function to deallocate this string.

Failure to call from_raw will lead to a memory leak.

pub fn into_string(self) -> Result<String, IntoStringError>

Converts the CString into a String if it contains valid Unicode data.

On failure, ownership of the original CString is returned.

pub fn into_bytes(self) -> Vec<u8>

Returns the underlying byte buffer.

The returned buffer does not contain the trailing nul separator and it is guaranteed to not have any interior nul bytes.

pub fn into_bytes_with_nul(self) -> Vec<u8>

Equivalent to the into_bytes function except that the returned vector includes the trailing nul byte.

pub fn as_bytes(&self) -> &[u8]

Returns the contents of this CString as a slice of bytes.

The returned slice does not contain the trailing nul separator and it is guaranteed to not have any interior nul bytes.

pub fn as_bytes_with_nul(&self) -> &[u8]

Equivalent to the as_bytes function except that the returned slice includes the trailing nul byte.

pub fn as_c_str(&self) -> &CStr

Extracts a CStr slice containing the entire string.

Important traits for Box<I>
impl<I> Iterator for Box<I> where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn into_boxed_c_str(self) -> Box<CStr>

Converts this CString into a boxed CStr.