Struct sgx_trts::enclave::SgxThreadData

pub struct SgxThreadData { /* fields omitted */ }

Methods

impl SgxThreadData

pub fn new() -> Self

get thread_data per thread.

Note

This API is only an experimental funtion.

pub fn td_base(&self) -> usize

td_base is to get TD base address per thread.

Note

This API is only an experimental funtion.

pub fn stack_base(&self) -> usize

stack_base is to get stack base address per thread.

Note

This API is only an experimental funtion.

pub fn stack_limit(&self) -> usize

stack_limit is to get stack limit per thread.

Note

This API is only an experimental funtion.

pub fn tls_base(&self) -> usize

tls_base is to get tls base address per thread.

Note

This API is only an experimental funtion.

pub fn last_error(&self) -> usize

last_error is to get last error per thread.

Note

This API is only an experimental funtion.

pub fn exception_flag(&self) -> usize

exception_flag is to get exception flag per thread.

Note

This API is only an experimental funtion.