Function sgx_trts::rsgx_raw_is_outside_enclave

pub fn rsgx_raw_is_outside_enclave(addr: *const u8, size: usize) -> bool

rsgx_raw_is_outside_enclave checks whether a given address is outside enclave memory.

The rsgx_raw_is_outside_enclave function checks that the buffer located at the pointer addr with its length of size is an address that is strictly outside the calling enclave address space.

Description

rsgx_raw_is_outside_enclave simply compares the start and end address of the buffer with the calling enclave address space. It does not check the property of the address.

Parameters

addr

The start address of the buffer.

size

The size of the buffer.

Requirements

Library: libsgx_trts.a

Return value

true

The buffer is strictly outside the enclave address space.

false

The whole buffer or part of the buffer is not outside the enclave, or the buffer is wrapped around.