Module sgx_trts::panicking

Implementation of various bits and pieces of the panic! macro and associated runtime pieces.

Structs

Location
A struct containing information about the location of a panic.

PanicInfo
A struct providing information about a panic.

Functions

begin_panic
This is the entry point of panicking for panic!() and assert!().

begin_panic_fmt
The entry point for panicking with a formatted message.

panicking
Determines whether the current thread is unwinding because of panic.

rust_begin_panic
Entry point of panic from the libcore crate.

rust_panic
A private no-mangle function on which to slap yer breakpoints.

set_panic_handler
Registers a custom panic handler, replacing any that was previously registered.

try⚠
Invoke a closure, capturing the cause of an unwinding panic if one occurs.

update_count_then_panic
Shim around rust_panic. Called by resume_unwind.

update_panic_count