Crate sgx_trts

Trusted Runtime System

The Intel(R) SGX trusted runtime system (tRTS) is a key component of the Intel(R) Software Guard Extensions SDK. It provides the enclave entry point logic as well as other functions to be used by enclave developers.

Intel(R) Software Guard Extensions Helper Functions

CustomExceptionHandling

Intel(R) Software Guard Extensions Helper Functions

The tRTS provides the helper functions for you to determine whether a given address is within or outside enclave memory.

The tRTS provides a wrapper to the RDRAND instruction to generate a true random number from hardware. enclave developers should use the rsgx_read_rand function to get true random numbers.

CustomExceptionHandling

The Intel(R) Software Guard Extensions SDK provides an API to allow you to register functions, or exception handlers, to handle a limited set of hardware exceptions. When one of the enclave supported hardware exceptions occurs within the enclave, the registered exception handlers will be called in a specific order until an exception handler reports that it has handled the exception. For example, issuing a CPUID instruction inside an Enclave will result in a #UD fault (Invalid Opcode Exception). ISV enclave code can call rsgx_register_exception_handler to register a function of type sgx_exception_handler_t to respond to this exception. To check a list of enclave supported exceptions, see Intel(R) Software Guard Extensions Programming Reference.

Note

Custom exception handling is only supported in HW mode. Although the exception handlers can be registered in simulation mode, the exceptions cannot be caught and handled within the enclave.

Note

OCALLs are not allowed in the exception handler.

Note

Custom exception handing only saves general purpose registers in sgx_ exception_info_t. You should be careful when touching other registers in the exception handlers.

Note

If the exception handlers can not handle the exceptions, abort() is called. abort() makes the enclave unusable and generates another exception.

Functions

rsgx_data_is_outside_enclave
rsgx_data_is_outside_enclave checks whether a given address is outside enclave memory.

rsgx_data_is_within_enclave
rsgx_data_is_within_enclave checks whether a given address is within enclave memory.

rsgx_get_enclave_base
rsgx_get_enclave_base is to get enclave image base address.

rsgx_get_heap_base
rsgx_get_heap_base is to get enclave heap base address.

rsgx_get_heap_size
rsgx_get_heap_size is to get enclave heap size.

rsgx_get_thread_data
rsgx_get_thread_data is to get TD base address per thread.

rsgx_raw_is_outside_enclave
rsgx_raw_is_outside_enclave checks whether a given address is outside enclave memory.

rsgx_raw_is_within_enclave
rsgx_raw_is_within_enclave checks whether a given address is within enclave memory.

rsgx_read_rand
rsgx_read_rand function is used to generate a random number inside the enclave.

rsgx_register_exception_handler
rsgx_register_exception_handler register an exception handler.

rsgx_slice_is_outside_enclave
rsgx_slice_is_outside_enclave checks whether a given address is outside enclave memory.

rsgx_slice_is_within_enclave
rsgx_slice_is_within_enclave checks whether a given address is within enclave memory.

rsgx_unregister_exception_handler
rsgx_unregister_exception_handler is used to unregister a custom exception handler.

Type Definitions

exception_handle