Function sgx_trts::veh::rsgx_unregister_exception_handler

pub fn rsgx_unregister_exception_handler(handle: exception_handle) -> bool

rsgx_unregister_exception_handler is used to unregister a custom exception handler.

Description

The Intel(R) SGX SDK supports the registration of custom exception handler functions. An enclave developer can write their own code to handle a limited set of hardware exceptions.

Calling rsgx_unregister_exception_handler allows developers to unregister an exception handler that was registered earlier.

Parameters

handle

A handle to the custom exception handler previously registered using the rsgx_register_exception_handler function.

Requirements

Library: libsgx_trts.a

Return value

true

The custom exception handler is unregistered successfully.

false

The exception handler was not unregistered (not a valid pointer, handler not found).